Move other directory utils
