+++ src/ExerciseList.hs
@@ -3,37 +3,7 @@ module ExerciseList where
 import qualified Data.Map            as M
 
 import           ExecutableExercises
-
--- There are three types of exercises.
--- 1. Some succeed once they have compiled (CompileOnly).
--- 2. Some require that unit tests pass on the exercise functions (UnitTests)
--- 3. Others involve some degree of IO. The User can use 'haskellings exec' on
---    these. There is one set of inputs and expected outputs as a quick test,
---    when the exercise is run via 'haskellings run' or the Watcher.
-data ExerciseType =
-  CompileOnly |
-  UnitTests |
-  -- One set of expected input lines and predicate for output lines.
-  Executable [String] ([String] -> Bool)
-
--- Manual instances needed because 'Executable' contains a function.
-instance Eq ExerciseType where
-  CompileOnly == CompileOnly = True
-  UnitTests == UnitTests = True
-  (Executable _ _) == (Executable _ _) = True
-  _ == _ = False
-
-instance Show ExerciseType where
-  show CompileOnly      = "CompileOnly"
-  show UnitTests        = "UnitTests"
-  show (Executable _ _) = "Executable"
-
-data ExerciseInfo = ExerciseInfo
-  { exerciseName      :: String
-  , exerciseDirectory :: String
-  , exerciseType      :: ExerciseType
-  , exerciseHint      :: String
-  } deriving (Show, Eq)
+import           Types
 
 allExercises :: [ExerciseInfo]
 allExercises =
